chatapp-backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
channels==4.1.0
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.1
ez_setup==0.9
mysqlclient==2.2.4
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
